refactor: cli enum
